Change WORKDIR to /workspace (#230)
The upstream oras image is using the /workspace directory as the default
working directory[1].
This is not the case in the Konflux rebuild where /home/oras is used
instead.

This can result in user confusion if they follow upstream oras docs
where commands assuming the usage of /workspace directory are used.

[1]: https://oras.land/docs/installation/#docker-image

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>